fix: FP32 exports + FSQ fallback + BPE path + TRT FP16 disable
- Export scripts: changed default dtype from float16 to float32.
  TRT handles FP16 optimization internally when building engines.
  FP16 ONNX models require FP16 input tensors from C++ which is
  inconvenient. FP32 ONNX lets TRT choose precision per layer.

- text-enc-ort.h: disabled trt_fp16_enable to prevent layernorm
  overflow (values up to 4.5 trillion from FP16 reduce/pow).

- pipeline-synth.cpp: FSQ tok/detok fallback to safetensors DiT dir
  when on ONNX path (FSQ not yet exported to ONNX).

- pipeline-synth-ops.cpp: resolve BPE path to parent directory when
  text_encoder_path points at an .onnx file.

+++ tools/onnx-export/export_cond_enc.py
@@ -206,7 +206,7 @@ class CondEncoderWrapperFixed(nn.Module):
         return enc_hidden
 
 
-def load_model(model_dir: str, device: str = "cuda", dtype=torch.float16):
+def load_model(model_dir: str, device: str = "cuda", dtype=torch.float32):
     """Load the AceStep model and extract the condition encoder."""
     model_dir = Path(model_dir)
